replace digital with macro

+++ Include/zstack/algo.h
@@ -1,10 +1,10 @@
 #ifndef ZSTACK_ALGO_H
 #define ZSTACK_ALGO_H
 
-extern void fft(double * data, int n, int isInverse);
+extern void fft(float * data, int n, int isInverse);
 extern int wavegen(double *output, int length, int freq, int sample_freq);
 extern int wavegen_complex(double *output, int length, int freq, int sample_freq);
-extern int hanning_window(double *io, unsigned int size);
-extern int linear_regression(double *y, unsigned int length);
+extern int hanning_window(float *io, unsigned int size);
+extern int linear_regression(float *y, unsigned int length);
 
 #endif
